Add files via upload
1) 3 octave piano -
For the first project, I made a 3 octave piano in which I have installed a PIR sensor so that whenever a person comes within the range, the LCD display will display a message 'Turn up the music' and an attached RGB LED will turn from red to green. I connected a potentiometer using which the user can choose one between three available octaves of low, medium and high frequencies. I attached 8 pushbuttons in order to be able to play all the notes of the respective chosen octave and a piezo is used to generate the sound. 
Link to the tinkercad circuit - https://www.tinkercad.com/things/4qR4ZYjam9G-3-octave-piano

2) Mars rover weather checker -  
I have incorporated a temperature sensor, a soil sensor and a photodiode in this circuit. It can be used in a mars rover. 
The temperature is detected by a temperature sensor [TMP36] and if it exceeds a certain limit - 90 degrees, a fan is turned on - I have represented the fan by a DC motor in this circuit. This can be used in real life applications to cool the internal components in the chassis in case the temperature gets too high. The photodiode reads the light intensity and both the temperature and the light intensity at each instant is displayed on the LCD. The soil sensor reads the moisture content from the soil and depending on whether the soil is dry, medium or wet, we rotate the servo motor accordingly by servo.write(0), servo.write(90) or servo.write(180). In real life, I'd attach an arrow on the servo and attach it to a display showing dry, medium and wet conditions so that we can see the soil moisture content through the rotation of the servo. 
In real life, I would additionally attach an SD card module in order to store the data. As tinkercad doesnt support SD cards, I couldnt execute it in the simulation. I would include an RTC module and an SD card module and write data to the SD card using a file object. This log consisting of daily temperature, light intensity and soil moisture conditions would help scientists analyse the weather of Mars effectively.
Link to the tinkercad circuit - https://www.tinkercad.com/things/dHou9ciloaN-mars-rover-weather-checker

3) Ultrasonic object detector - 
In this project, I have a servo motor which rotates upto its full capacity (in tinkercad, it can only go from 0 - 270 degrees, In reality, it can go from 0 - 360 degrees). In real life, I will mount the ultrasonic sensor on top of the servo motor so it rotates and senses if there's an object present in the particular specified range and if it is present the LED present in that particular direction will glow. It works on the functioning of a radar. Through this we'll know the approximate direction of the object through the corresponding glowing LEDs.
Link to the tinkercad circuit - https://www.tinkercad.com/things/8qomOeF1aAc-ultrasonic-object-detector